+++ constant/queryKey.ts
@@ -1,6 +1,4 @@
-import { QueryKeyType } from '@/types/queryKeyType';
-
-export const QUERY_KEY: QueryKeyType = {
+export const QUERY_KEY = {
   ITEM_DATA: 'itemData',
   WISHES_DATA: 'wishesData',
   PAYREADY: 'payReady',
@@ -12,4 +10,5 @@ export const QUERY_KEY: QueryKeyType = {
   CAKE_LETTERS: 'cakeLetters',
   WISH_LINKS: 'wishLinks',
   ONE_WISH: 'oneWish',
+  PRE_SIGNED_URL: 'preSignedURL',
 };
