+++ hooks/queries/wishes/useCreateWishesLink.ts
@@ -8,8 +8,6 @@ export function useCreateWishesLink() {
   const wishesData = useRecoilValue(WishesData);
   const setLoginUserInfo = useSetRecoilState(LoginUserInfo);
 
-  console.log(wishesData);
-
   const { mutate: postWishesData, isSuccess } = useMutation(() => createWishesLink(wishesData), {
     onSuccess: (data) => {
       console.log(data);
