--- hooks/wishes/useUploadItemInfo.tsx
@@ -1,21 +1,0 @@
-import { useState } from 'react';
-
-export default function useUploadItemInfo() {
-  const [imageFile, setImageFile] = useState<File | Blob | null>(null);
-  const [previewImage, setPreviewImage] = useState<string | ArrayBuffer | null>('');
-
-  function uploadImageFile(e: React.ChangeEvent<HTMLInputElement>) {
-    const imageFile = e.target.files && e.target.files[0];
-
-    if (imageFile) {
-      setImageFile(imageFile);
-      const reader = new FileReader();
-      imageFile && reader.readAsDataURL(imageFile);
-      reader.onloadend = () => {
-        setPreviewImage(reader.result);
-      };
-    }
-  }
-
-  return { imageFile, previewImage, setPreviewImage, uploadImageFile };
-}
