--- types/queryKeyType.ts
@@ -1,13 +1,0 @@
-export interface QueryKeyType {
-  ITEM_DATA: string;
-  WISHES_DATA: string;
-  PAYREADY: string;
-  PG_TOKEN: string;
-  USER: string;
-  ACCOUNT: string;
-  PROGRESS: string;
-  CAKES_COUNT: string;
-  CAKE_LETTERS: string;
-  WISH_LINKS: string;
-  ONE_WISH: string;
-}
